Fixed remaining clang compiler warnings
https://codereview.appspot.com/6865053/



git-svn-id: http://skia.googlecode.com/svn/trunk@6646 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/effects/SkColorMatrixFilter.cpp
@@ -417,7 +417,7 @@ GrEffect* ColorMatrixEffect::TestCreate(SkRandom* random,
                                         GrContext*,
                                         GrTexture* dummyTextures[2]) {
     SkColorMatrix colorMatrix;
-    for (int i = 0; i < SK_ARRAY_COUNT(colorMatrix.fMat); ++i) {
+    for (size_t i = 0; i < SK_ARRAY_COUNT(colorMatrix.fMat); ++i) {
         colorMatrix.fMat[i] = random->nextSScalar1();
     }
     return SkNEW_ARGS(ColorMatrixEffect, (colorMatrix));
